requirements.txt
Django>=3.1.7,<3.2.0
djangorestframework>=3.12.2,<3.13.0
psycopg2>=2.7.5,<2.8.0

flake8>=3.8.4,<3.9.0